# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WAYAPP WhatsApp Gateway REST API",
    "version": "1.0.0",
    "description": "Enterprise Meta Cloud API Gateway & Marketing Automation Platform REST API v1"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1 Endpoint"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { "ApiKeyAuth": [] },
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/campaigns": {
      "get": {
        "summary": "List Campaigns",
        "responses": { "200": { "description": "Campaign list" } }
      },
      "post": {
        "summary": "Create Broadcast Campaign",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Ramadan Special Promo" },
                  "templateId": { "type": "string" },
                  "audienceFilter": { "type": "object" },
                  "variableMappings": { "type": "object" },
                  "startImmediately": { "type": "boolean", "default": false }
                },
                "required": ["name", "templateId"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Campaign created" } }
      }
    },
...
  }
}